Move the argnames checks to Lambda::new, enforce it as class invariant
Sadly I can't fully enforce this as a class invariant because List is made up of Rc<RefCell<List>>.
There's probably a reason for that but I can't find it.

This is technically a breaking change, as argnames, which used to be a `pub Rc<Value>`, is now a `priv Rc<List>`.
It can still be accessed through the `argnames(&self)` method, which can be used as a drop-in replacement to
`lambda.argnames.as_list().unwrap()`.